telegramstream/requirements.txt
telethon>=1.35.0
pyyaml>=6.0
prompt-toolkit>=3.0.30
nats-py>=2.7.0
emoji>=2.0.0
regex>=2023.5.0
aiohttp>=3.9.0